Added label 'id' to job metrics.
Previously we used the cluster id to identify a job cluster, however this value
is unique only at the machine level, two jobs executed by different machines on the pool
may have the same cluster id.
Here we add a unique identifier obtained from concatenating the cluster id and the submmiter name
in the form '3@machine'.
The Job Details dashboard was changed accordingly to now used the new id to identify clusters.